requirements.txt
fastapi==0.111.0
uvicorn==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.32
pydantic==2.7.4
pydantic-settings==2.4.0
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
PyMySQL==1.1.0
streamlit==1.37.1
requests==2.32.3